+++ lib/features/immersion/parameters/immersion_filtres_middleware.dart
@@ -1,10 +1,11 @@
 import 'package:pass_emploi_app/features/immersion/list/immersion_list_actions.dart';
-import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_actions.dart';
-import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_state.dart';
 import 'package:pass_emploi_app/features/login/login_state.dart';
 import 'package:pass_emploi_app/redux/app_state.dart';
 import 'package:pass_emploi_app/repositories/immersion_repository.dart';
 import 'package:redux/redux.dart';
+
+import 'immersion_search_parameters_actions.dart';
+import 'immersion_search_parameters_state.dart';
 
 class ImmersionFiltresMiddleware extends MiddlewareClass<AppState> {
   final ImmersionRepository _repository;
