+++ lib/presentation/immersion/immersion_filtres_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:equatable/equatable.dart';
 import 'package:pass_emploi_app/features/immersion/list/immersion_list_state.dart';
-import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_actions.dart';
-import 'package:pass_emploi_app/features/immersion/search/immersion_search_parameters_state.dart';
+import 'package:pass_emploi_app/features/immersion/parameters/immersion_search_parameters_actions.dart';
+import 'package:pass_emploi_app/features/immersion/parameters/immersion_search_parameters_state.dart';
 import 'package:pass_emploi_app/models/immersion_filtres_parameters.dart';
 import 'package:pass_emploi_app/presentation/display_state.dart';
 import 'package:pass_emploi_app/redux/app_state.dart';
