+++ lib/redux/store_factory.dart
@@ -11,7 +11,7 @@ import 'package:pass_emploi_app/features/favori/update/data_from_id_extractor.da
 import 'package:pass_emploi_app/features/favori/update/favori_update_middleware.dart';
 import 'package:pass_emploi_app/features/immersion/details/immersion_details_middleware.dart';
 import 'package:pass_emploi_app/features/immersion/list/immersion_list_middleware.dart';
-import 'package:pass_emploi_app/features/immersion/search/immersion_filtres_middleware.dart';
+import 'package:pass_emploi_app/features/immersion/parameters/immersion_filtres_middleware.dart';
 import 'package:pass_emploi_app/features/location/search_location_middleware.dart';
 import 'package:pass_emploi_app/features/login/login_middleware.dart';
 import 'package:pass_emploi_app/features/metier/search_metier_middleware.dart';
